requirements.txt
asgiref==3.9.1
backports.tarfile==1.2.0
bcrypt==4.0.1
build==1.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
crispy-bootstrap5==2025.6
cryptography==44.0.1
defusedxml==0.7.1
-e git+https://github.com/SamCzarski/DevWorksDjangoLib.git@9c7b21a58ebdfcdd89acb272a1678448358651a8#egg=DevWorksDjangoLib
Django==5.2.5
django-cors-headers==4.3.1
django-crispy-forms==2.4
django-csp==3.7
django-encrypted-model-fields==0.6.5
django-extensions==3.2.3
django-filter==25.1
django-log-request-id==2.1.0
django-password-validators==1.7.3
django-ratelimit==4.1.0
django-redis==6.0.0
django-simple-history==3.10.1
djangorestframework==3.15.2
djangorestframework-csv==3.0.2
docutils==0.22
gunicorn==23.0.0
id==1.5.0
idna==3.10
importlib_metadata==8.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jeepney==0.9.0
jwcrypto==1.5.6
keyring==25.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.8.0
nh3==0.3.0
oauthlib==3.3.1
packaging==25.0
PGPy==0.6.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydash==7.0.6
Pygments==2.19.2
PyJWT==2.8.0
pyOpenSSL==25.1.0
pyproject_hooks==1.2.0
python-json-logger==3.3.0
python3-openid==3.2.0
PyYAML==6.0.2
readme_renderer==44.0
redis==6.4.0
requests==2.32.2
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.1.0
SecretStorage==3.4.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
twine==6.2.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
